+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update design of audio player to remove need for settings panel.
 - Download rather than stream audio.
 - Start playing audio from current verse.
+- Play verse when verse number tapped (if audio player is visible).
 
 ## 0.14.0 - January 10, 2025